+++ tests/e2e/Services/Databases/DatabasesConsoleClientTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\E2E\Services\Database;
+namespace Tests\E2E\Services\Databases;
 
 use Tests\E2E\Scopes\Scope;
 use Tests\E2E\Scopes\ProjectCustom;
